fix: OAuth 에러 핸들링 추가

+++ client/src/pages/OAuth.jsx
@@ -2,17 +2,33 @@ import { useSetRecoilState } from 'recoil';
 import Profile from '../recoil/profile';
 import { useNavigate } from 'react-router-dom';
 import { useEffect } from 'react';
+import ModalState from '../recoil/modal';
 
 const OAuth = () => {
   const setProfile = useSetRecoilState(Profile);
+  const setModal = useSetRecoilState(ModalState);
   const navigate = useNavigate();
 
   const urlParams = new URL(location.href).searchParams;
   const Authorization = urlParams.get('Authorization');
   const userStatus = urlParams.get('userStatus');
   const userId = urlParams.get('userId');
+  const code = urlParams.get('code');
+  const message = urlParams.get('message');
 
   useEffect(() => {
+    if (code === '409' && message === 'USER_EMAIL_EXISTS') {
+      navigate('/login');
+      setModal({
+        isOpen: true,
+        modalType: 'alert',
+        props: {
+          text: '이미 일반회원으로 가입된 계정입니다.\n일반 로그인을 이용해주세요.',
+        },
+      });
+      return;
+    }
+
     sessionStorage.setItem('authorization', Authorization);
     sessionStorage.setItem('userId', userId);
     sessionStorage.setItem('userStatus', userStatus);
